+++ src/navigation/AppNavigator.tsx
@@ -5,6 +5,7 @@ import HomeScreen from '../presentation/screens/HomeScreen';
 import AddRecordScreen from '../presentation/screens/AddRecordScreen';
 import DetailScreen from '../presentation/screens/DetailScreen';
 import ExportScreen from '../presentation/screens/ExportScreen';
+import FlightRecordsScreen from '../presentation/screens/FlightRecordsScreen'; // 追加
 // import InspectionRecordScreen from '../presentation/screens/InspectionRecordScreen'; // 後で追加予定
 // import MaintenanceRecordScreen from '../presentation/screens/MaintenanceRecordScreen'; // 後で追加予定
 
@@ -13,6 +14,7 @@ export type RootStackParamList = {
   AddRecord: undefined;
   Detail: { record: any };
   Export: undefined;
+  FlightRecords: undefined;
   // InspectionRecord: undefined; // 後で追加予定
   // MaintenanceRecord: undefined; // 後で追加予定
 };
@@ -24,12 +26,10 @@ const AppNavigator = () => {
     <NavigationContainer>
       <Stack.Navigator initialRouteName="Home">
         <Stack.Screen name="Home" component={HomeScreen} />
+        <Stack.Screen name="FlightRecords" component={FlightRecordsScreen} /> 
         <Stack.Screen name="AddRecord" component={AddRecordScreen} />
         <Stack.Screen name="Detail" component={DetailScreen} />
         <Stack.Screen name="Export" component={ExportScreen} />
-        {/* 後で追加予定 */}
-        {/* <Stack.Screen name="InspectionRecord" component={InspectionRecordScreen} /> */}
-        {/* <Stack.Screen name="MaintenanceRecord" component={MaintenanceRecordScreen} /> */}
       </Stack.Navigator>
     </NavigationContainer>
   );
